Shift findEnrolmentFromList method and tests to ModelManager
