+++ src/test/java/seedu/ccacommander/model/ModelManagerTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.ccacommander.model.Model.PREDICATE_SHOW_ALL_EVENTS;
 import static seedu.ccacommander.model.Model.PREDICATE_SHOW_ALL_MEMBERS;
+import static seedu.ccacommander.model.ModelManager.findEnrolmentFromList;
 import static seedu.ccacommander.testutil.Assert.assertThrows;
 import static seedu.ccacommander.testutil.TypicalEvents.AURORA_BOREALIS;
 import static seedu.ccacommander.testutil.TypicalEvents.BOXING_DAY;
@@ -13,18 +14,24 @@ import static seedu.ccacommander.testutil.TypicalMembers.BENSON;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
 import seedu.ccacommander.commons.core.GuiSettings;
 import seedu.ccacommander.model.enrolment.Enrolment;
+import seedu.ccacommander.model.enrolment.exceptions.EnrolmentNotFoundException;
 import seedu.ccacommander.model.event.EventNameContainsKeywordsPredicate;
 import seedu.ccacommander.model.exceptions.RedoStateException;
 import seedu.ccacommander.model.exceptions.UndoStateException;
 import seedu.ccacommander.model.member.MemberNameContainsKeywordsPredicate;
 import seedu.ccacommander.testutil.CcaCommanderBuilder;
 import seedu.ccacommander.testutil.EnrolmentBuilder;
+import seedu.ccacommander.testutil.TypicalEnrolments;
+import seedu.ccacommander.testutil.TypicalEvents;
+import seedu.ccacommander.testutil.TypicalMembers;
 
 public class ModelManagerTest {
 
@@ -202,6 +209,39 @@ public class ModelManagerTest {
         modelManager.createEnrolment(enrolment);
         modelManager.deleteEnrolment(enrolment);
         assertFalse(modelManager.hasEvent(AURORA_BOREALIS));
+    }
+
+    @Test
+    public void findEnrolmentFromList_success() {
+        List<Enrolment> enrolmentList = new ArrayList<>();
+
+        Enrolment enrolmentToFindOne = TypicalEnrolments.ALICE_AURORA;
+        Enrolment enrolmentToFindTwo = TypicalEnrolments.GEORGE_GRAVITY_DISCOVERY_DAY;
+
+        enrolmentList.add(enrolmentToFindOne);
+        enrolmentList.add(enrolmentToFindTwo);
+
+        assertEquals(enrolmentToFindOne, findEnrolmentFromList(enrolmentList,
+                TypicalMembers.ALICE.getName(), TypicalEvents.AURORA_BOREALIS.getName()));
+        assertEquals(enrolmentToFindTwo, findEnrolmentFromList(enrolmentList,
+                TypicalMembers.GEORGE.getName(), TypicalEvents.GRAVITY_DISCOVERY_DAY.getName()));
+    }
+
+    @Test
+    public void findEnrolmentFromList_fail() {
+        List<Enrolment> enrolmentList = new ArrayList<>();
+
+        Enrolment enrolmentToFindOne = TypicalEnrolments.ALICE_AURORA;
+        Enrolment enrolmentToFindTwo = TypicalEnrolments.GEORGE_GRAVITY_DISCOVERY_DAY;
+
+        enrolmentList.add(enrolmentToFindOne);
+        enrolmentList.add(enrolmentToFindTwo);
+
+        try {
+            findEnrolmentFromList(enrolmentList, TypicalMembers.CARL.getName(), TypicalEvents.BOXING_DAY.getName());
+        } catch (EnrolmentNotFoundException ee) {
+            assertEquals("Enrolment cannot be found", ee.getMessage());
+        }
     }
 
     @Test
